requirements.txt
numpy==1.24.4
streamlit==1.45.1
cohere==5.15.0
scikit-learn==1.6.1
pandas==2.2.2
PyPDF2==3.0.1
gTTS==2.5.1
speechrecognition==3.10.0
streamlit-webrtc==0.45.2
sounddevice==0.4.6
soundfile==0.12.1
python-dotenv==1.0.1
torchaudio==2.2.1
torch==2.2.1
git+https://github.com/openai/whisper.git

